+++ test/functional/feature_dersig.py
@@ -103,7 +103,10 @@ class BIP66Test(BitcoinTestFramework):
         block.solve()
 
         with self.nodes[0].assert_debug_log(
-            expected_msgs=[f"ConnectBlock {block.hash_hex} failed, blk-bad-inputs"]
+            expected_msgs=[
+                f"BlockChecked: block hash={block.hash_hex} "
+                f"state=mandatory-script-verify-flag-failed (Non-canonical DER signature)"
+            ]
         ):
             peer.send_and_ping(msg_block(block))
             assert_equal(self.nodes[0].getbestblockhash(), tip)